UnitRec coef as a Rational

+++ src/main/scala/com/manyangled/coulomb/Quantity.scala
@@ -335,5 +335,5 @@ object ConvertableTemps {
   implicit def witnessConvertableTemps[U1 <: TemperatureExpr, U2 <: TemperatureExpr](implicit
       turecU1: TempUnitRec[U1], urecU1: UnitRec[U1],
       turecU2: TempUnitRec[U2], urecU2: UnitRec[U2]): ConvertableTemps[U1, U2] =
-    new ConvertableTemps[U1, U2](urecU1.coef, turecU1.offset, urecU2.coef, turecU2.offset)
+    new ConvertableTemps[U1, U2](urecU1.coef.toDouble, turecU1.offset, urecU2.coef.toDouble, turecU2.offset)
 }
